созданы дополнительно два js файла и код распределён между файлами
